# backend/candidate/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.CandidateCreateRequest:
    properties:
      birthday:
        type: string
      city:
        type: string
      phone:
        type: string
      telegram:
        type: string
    type: object
  github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.CandidateResponse:
    properties:
      birthday:
        type: string
      city:
        type: string
      id:
        type: string
      phone:
        type: string
      telegram:
        type: string
      user_id:
        type: string
    type: object
  github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.CandidateUpdateRequest:
    properties:
      birthday:
        type: string
      city:
        type: string
      phone:
        type: string
      telegram:
        type: string
    type: object
  github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.CreateResumeRequest:
    properties:
      data:
        $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ResumeData'
      name:
        type: string
      status:
        type: integer
    type: object
  github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.Education:
    properties:
      end_year:
        type: integer
      faculty:
        type: string
      format:
        type: string
      level:
        type: string
      specialization:
        type: string
      start_year:
        type: integer
      university:
        type: string
    type: object
  github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ResumeData:
    properties:
      additional_info:
        type: string
      citizenship:
        type: string
      city:
        type: string
      date_of_birth:
        type: string
      desired_format:
        type: string
      education:
        items:
          $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.Education'
        type: array
      email:
        type: string
      english_level:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone:
        type: string
      portfolio_link:
        type: string
      skills_list:
        description: Store skill IDs
        items:
          type: string
        type: array
      work_experiences:
        items:
          $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.WorkExperience'
        type: array
    type: object
  github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ResumeResponse:
    properties:
      candidate_id:
        type: string
      created_at:
        type: string
      data:
        $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ResumeData'
      id:
        type: string
      name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
  github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.SkillResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.UpdateResumeRequest:
    properties:
      data:
        $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ResumeData'
      id:
        type: string
      name:
        type: string
      status:
        type: integer
    type: object
  github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.WorkExperience:
    properties:
      company:
        type: string
      period:
        type: string
      position:
        type: string
      responsibilities:
        type: string
    type: object
host: 0.0.0.0:8081
info:
  contact: {}
  description: API for managing candidate profiles
  title: 'Trainee Match: Candidate Service API'
  version: "1.0"
paths:
  /candidate/:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Candidate creation data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.CandidateUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.CandidateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
      summary: Update candidate profile
      tags:
      - candidate
    post:
      consumes:
      - application/json
      description: Creates a new candidate profile associated with the authenticated
        user
      parameters:
      - description: Candidate creation data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.CandidateCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.CandidateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
      summary: Create candidate profile
      tags:
      - candidate
  /candidate/me:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.CandidateResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
      summary: Get my candidate profile
      tags:
      - candidate
  /resume/:
    post:
      consumes:
      - application/json
      parameters:
      - description: Resume creation data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.CreateResumeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ResumeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
      summary: Create a new resume
      tags:
      - resume
  /resume/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Resume ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ResumeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
      summary: Get resume by ID
      tags:
      - resume
    patch:
      consumes:
      - application/json
      parameters:
      - description: Resume ID
        in: path
        name: id
        required: true
        type: string
      - description: Resume update data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.UpdateResumeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ResumeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
      summary: Update a resume
      tags:
      - resume
  /skill/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Skill ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.SkillResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
      summary: Get skill by ID
      tags:
      - skill
  /skill/list:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.SkillResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_HghaVlad_trainee-match_backend_candidate_internal_delivery_http_dto.ErrorResponse'
      summary: List all skills
      tags:
      - skill
  /swagger/:
    get:
      consumes:
      - application/json
      description: This endpoint serves the Swagger UI for the API documentation.
      produces:
      - text/html
      responses:
        "200":
          description: Swagger UI served successfully
          schema:
            type: string
      summary: Swagger UI
      tags:
      - swagger
schemes:
- http
- https
swagger: "2.0"